+++ lib/CacheAsset.js
