+++ lib/CacheEnhancedResolve.js
