+++ lib/CacheModule.js
