+++ lib/CacheModuleResolver.js
